requirements.txt
rumps>=0.4.0
pyobjc-framework-Cocoa>=9.0
pyinstaller>=6.0
pytest>=7.0
